Disable RC4 Ciphers
The RC4 algorithm, as used in the TLS protocol and SSL protocol, does not properly combine state data with key data during the initialization phase. This makes it easier for remote attackers to conduct plaintext-recovery attacks against the initial bytes of a stream by sniffing network traffic that occasionally relies on keys affected by the Invariance Weakness, and then using a brute-force approach involving LSB values. This is also known as the "Bar Mitzvah" issue.

A successful attack against a TLS/SSL session that uses the RC4 cipher could result in a leak of a limited amount of plaintext such as user credentials.