Create MS19-NOV registry.xml
By default, all Windows client operating systems and Windows Server 2019 have the mitigation for this vulnerability enabled. However, the earlier Windows Server versions do not have this mitigation enabled by default.

An information disclosure vulnerability exists when the Windows kernel improperly handles objects in memory. To exploit this vulnerability, an attacker would have to log on to an affected system and run a specially crafted application. The vulnerability would not allow an attacker to execute code or to elevate user rights directly, but it could be used to obtain information that could be used to try to further compromise the affected system.